powerpc: Introduce bigendian tune and use it by default
In prepare of powerpc64le support introduce new tune "bigendian" that is
on by default for all powerpc machines. For little-endian we will not
add this tune and add '-mlittle-endian' compiler option when "bigendian"
tune is not in TUNE_FEATURES.

Add "bigendian" to TUNE_FEATURES_tune-powerpc{,64} and use them in more
specific TUNE_FEATURES for machine whenever possible. This inherits base
powerpc machines features from architecture definitions.

Not adding TUNECONFLICTS for "m32" to give someone a chance to try it
and make working if necessary (may be at least for multilib case?).

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>