glibc: Rename ldd in multilib package
This is symmetrical to musl change that renames ldd to ldd.${PN} for
multilib package to avoid filename collision on install.

While there add ldd to glibc package RRECOMMENDS since it is a shell
script not larger than 6Kb in size that costs nothing in compare with
overall glibc binary size and utility could be useful for trivial
debugging steps. If required recommendation can be easily dropped
(e.g. via NO_RECOMMENDATIONS).

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>