qemuppc64: Add qemu machine for PowerPC64
Now we have working qemu-system-ppc64 build in qemu-native add qemuppc64
machine definition to make possible to runqemu on just built PowerPC64
image.

Use pSeries as qemu -machine and POWER8 as -cpu and build code for
POWER5 by default to make it run on KVM-PR on PowerMac G5, KVM-HV on
modern IBM POWER8 and later server class machines, as well as using
full emulation TCG accelerator. Use -machine features syntax to add
accelerators with fallback to TCG.

Another reason to use pSeries machine is that allows autoselection of
KVM accelerator type PR or HV depending on hardware capabilities and
loaded modules.

Boot tested on PowerMac G5 with KVM-PR and IBM POWER8 with KVM-HV as
well as with TCG on x86_64 host. Note that code build for e6500 and
e5500 boots correctly on IBM POWER8 machine with KVM-HV, but fails in
TCG mode.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>