qemuppc: Use KVM accelerator when available with fallback to TCG
Use -machine features syntax to provide list of accelerators we want to
try in order of preference when running PowerPC code with runqemu.

Note that KVM accelerator tried by default, if fails TCG used. There is
no need to pass kvm option to runqemu (and it will fail because only x86
virtualization is checked).

Boot tested on PowerMac G5 with KVM-PR and IBM POWER8 with KVM-HV as
well as with TCG on x86_64.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>